+++ TinyNvidiaUpdateChecker/Forms/LocationChooserForm.Designer.cs
@@ -28,77 +28,64 @@
         /// </summary>
         private void InitializeComponent()
         {
-            this.locationListView = new System.Windows.Forms.ListView();
-            this.columnLocation = ((System.Windows.Forms.ColumnHeader)(new System.Windows.Forms.ColumnHeader()));
-            this.ConfirmBtn = new System.Windows.Forms.Button();
-            this.label1 = new System.Windows.Forms.Label();
-            this.SuspendLayout();
-            // 
-            // locationListView
-            // 
-            this.locationListView.Columns.AddRange(new System.Windows.Forms.ColumnHeader[] {
-            this.columnLocation});
-            this.locationListView.FullRowSelect = true;
-            this.locationListView.GridLines = true;
-            this.locationListView.HideSelection = false;
-            this.locationListView.Location = new System.Drawing.Point(12, 51);
-            this.locationListView.MultiSelect = false;
-            this.locationListView.Name = "locationListView";
-            this.locationListView.Size = new System.Drawing.Size(217, 238);
-            this.locationListView.TabIndex = 0;
-            this.locationListView.UseCompatibleStateImageBehavior = false;
-            this.locationListView.View = System.Windows.Forms.View.List;
-            // 
-            // columnLocation
-            // 
-            this.columnLocation.Text = "Location";
-            this.columnLocation.Width = 347;
+            ConfirmBtn = new System.Windows.Forms.Button();
+            label1 = new System.Windows.Forms.Label();
+            comboBox = new System.Windows.Forms.ComboBox();
+            SuspendLayout();
             // 
             // ConfirmBtn
             // 
-            this.ConfirmBtn.Location = new System.Drawing.Point(12, 295);
-            this.ConfirmBtn.Name = "ConfirmBtn";
-            this.ConfirmBtn.Size = new System.Drawing.Size(75, 23);
-            this.ConfirmBtn.TabIndex = 1;
-            this.ConfirmBtn.Text = "Confirm";
-            this.ConfirmBtn.UseVisualStyleBackColor = true;
-            this.ConfirmBtn.Click += new System.EventHandler(this.ConfirmBtn_Click);
+            ConfirmBtn.Enabled = false;
+            ConfirmBtn.Location = new System.Drawing.Point(221, 67);
+            ConfirmBtn.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
+            ConfirmBtn.Name = "ConfirmBtn";
+            ConfirmBtn.Size = new System.Drawing.Size(75, 29);
+            ConfirmBtn.TabIndex = 1;
+            ConfirmBtn.Text = "Confirm";
+            ConfirmBtn.UseVisualStyleBackColor = true;
+            ConfirmBtn.Click += ConfirmBtn_Click;
             // 
             // label1
             // 
-            this.label1.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
-            | System.Windows.Forms.AnchorStyles.Right)));
-            this.label1.AutoSize = true;
-            this.label1.Location = new System.Drawing.Point(34, 19);
-            this.label1.Name = "label1";
-            this.label1.Size = new System.Drawing.Size(173, 17);
-            this.label1.TabIndex = 2;
-            this.label1.Text = "Choose download location";
+            label1.Anchor = System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left | System.Windows.Forms.AnchorStyles.Right;
+            label1.AutoSize = true;
+            label1.Location = new System.Drawing.Point(66, 25);
+            label1.Name = "label1";
+            label1.Size = new System.Drawing.Size(187, 20);
+            label1.TabIndex = 2;
+            label1.Text = "Choose download location";
+            // 
+            // comboBox
+            // 
+            comboBox.DropDownStyle = System.Windows.Forms.ComboBoxStyle.DropDownList;
+            comboBox.FormattingEnabled = true;
+            comboBox.Location = new System.Drawing.Point(12, 68);
+            comboBox.Name = "comboBox";
+            comboBox.Size = new System.Drawing.Size(187, 28);
+            comboBox.TabIndex = 3;
+            comboBox.SelectedIndexChanged += comboBox_SelectedIndexChanged;
             // 
             // LocationChooserForm
             // 
-            this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
-            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(241, 330);
-            this.Controls.Add(this.label1);
-            this.Controls.Add(this.ConfirmBtn);
-            this.Controls.Add(this.locationListView);
-            this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.None;
-            this.Icon = System.Drawing.Icon.ExtractAssociatedIcon(System.Windows.Forms.Application.Execut‌​ablePath);
-            this.Name = "LocationChooserForm";
-            this.StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;
-            this.Text = "TinyNvidiaUpdateChecker - Choose Download Location";
-            this.Load += new System.EventHandler(this.LocationChooserForm_Load);
-            this.ResumeLayout(false);
-            this.PerformLayout();
-
+            AutoScaleDimensions = new System.Drawing.SizeF(8F, 20F);
+            AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            ClientSize = new System.Drawing.Size(308, 112);
+            Controls.Add(comboBox);
+            Controls.Add(label1);
+            Controls.Add(ConfirmBtn);
+            FormBorderStyle = System.Windows.Forms.FormBorderStyle.None;
+            Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
+            Name = "LocationChooserForm";
+            StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;
+            Text = "TinyNvidiaUpdateChecker - Choose Download Location";
+            Load += LocationChooserForm_Load;
+            ResumeLayout(false);
+            PerformLayout();
         }
 
         #endregion
-
-        private System.Windows.Forms.ListView locationListView;
-        private System.Windows.Forms.ColumnHeader columnLocation;
         private System.Windows.Forms.Button ConfirmBtn;
         private System.Windows.Forms.Label label1;
+        private System.Windows.Forms.ComboBox comboBox;
     }
 }