+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Internet connection failing whilst downloading driver is now properly communicated, see issue #204
 ### Changed
+- Redesigned Download Location form
 - Updated .NET Runtime to version 8
 - Replaced `Microsoft.Windows.Compatibility` package with `System.Management`